+++ UnitTests/UnitTests.HighPerformance.Shared/Extensions/Test_ReadOnlySpanExtensions.Count.cs
@@ -5,10 +5,10 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics.Contracts;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using Microsoft.Toolkit.HighPerformance.Extensions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using UnitTests.HighPerformance.Shared.Buffers.Internals;
 
 #nullable enable
 
@@ -258,54 +258,6 @@ namespace UnitTests.HighPerformance.Extensions
             data.Span.Fill(value);
 
             return data;
-        }
-
-        /// <summary>
-        /// An owner for a buffer of an unmanaged type, recycling <see cref="byte"/> arrays to save memory.
-        /// </summary>
-        /// <typeparam name="T">The type of items to store in the rented buffers.</typeparam>
-        private sealed unsafe class UnmanagedSpanOwner<T> : IDisposable
-            where T : unmanaged
-        {
-            /// <summary>
-            /// The size of the current instance
-            /// </summary>
-            private readonly int length;
-
-            /// <summary>
-            /// The pointer to the underlying <see cref="byte"/> array.
-            /// </summary>
-            private IntPtr ptr;
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="UnmanagedSpanOwner{T}"/> class.
-            /// </summary>
-            /// <param name="size">The size of the buffer to rent.</param>
-            public UnmanagedSpanOwner(int size)
-            {
-                this.ptr = Marshal.AllocHGlobal(size * Unsafe.SizeOf<T>());
-                this.length = size;
-            }
-
-            /// <inheritdoc/>
-            public void Dispose()
-            {
-                IntPtr ptr = this.ptr;
-
-                if (ptr == IntPtr.Zero)
-                {
-                    return;
-                }
-
-                this.ptr = IntPtr.Zero;
-
-                Marshal.FreeHGlobal(ptr);
-            }
-
-            /// <summary>
-            /// Gets the <see cref="Memory{T}"/> for the current instance.
-            /// </summary>
-            public Span<T> Span => new Span<T>((void*)this.ptr, this.length);
         }
     }
 }
